refactoring fifo/wsgi servers into their own separate module
