+++ ai2thor/server.py
@@ -6,56 +6,13 @@ Handles all communication with Unity through a Flask service.  Messages
 are sent to the controller using a pair of request/response queues.
 """
 
-
-import json
-import logging
-import threading
-import sys
-import msgpack
-import os
-import os.path
-import tempfile
-from enum import IntEnum, unique
-from collections import defaultdict
-import struct
-
-try:
-    from queue import Queue, Empty
-except ImportError:
-    from Queue import Queue, Empty
-
-import time
 import warnings
-
-from flask import Flask, request, make_response, abort
-import werkzeug
-import werkzeug.serving
-import werkzeug.http
 import numpy as np
 from enum import Enum
-
 from ai2thor.util.depth import apply_real_noise, generate_noise_indices
+import json
+import sys
 
-logging.getLogger('werkzeug').setLevel(logging.ERROR)
-
-werkzeug.serving.WSGIRequestHandler.protocol_version = 'HTTP/1.1'
-
-MAX_DEPTH = 5000
-
-# get with timeout to allow quit
-def queue_get(que, unity_proc=None):
-    res = None
-    while True:
-        try:
-            res = que.get(block=True, timeout=0.5)
-            break
-        except Empty:
-            # we poll here for the unity proc in the event that it has
-            # exited otherwise we would wait indefinetly for the queue
-            if unity_proc and unity_proc.poll() is not None:
-                raise Exception("Unity process exited %s" % unity_proc.returncode)
-            pass
-    return res
 
 class NumpyAwareEncoder(json.JSONEncoder):
 
@@ -64,31 +21,6 @@ class NumpyAwareEncoder(json.JSONEncoder):
             return np.asscalar(obj)
         return super(NumpyAwareEncoder, self).default(obj)
 
-class BufferedIO(object):
-    def __init__(self, wfile):
-        self.wfile = wfile
-        self.data = []
-
-    def write(self, output):
-        self.data.append(output)
-
-    def flush(self):
-        self.wfile.write(b"".join(self.data))
-        self.wfile.flush()
-    def close(self):
-        return self.wfile.close()
-
-    @property
-    def closed(self):
-        return self.wfile.closed
-
-class ThorRequestHandler(werkzeug.serving.WSGIRequestHandler):
-    def run_wsgi(self):
-        old_wfile = self.wfile
-        self.wfile = BufferedIO(self.wfile)
-        result = super(ThorRequestHandler, self).run_wsgi()
-        self.wfile = old_wfile
-        return result
 
 class MultiAgentEvent(object):
 
@@ -347,62 +279,6 @@ class Event(object):
                 return obj
         return None
 
-
-class MultipartFormParser(object):
-
-    @staticmethod
-    def get_boundary(request_headers):
-        for h, value in request_headers:
-            if h == 'Content-Type':
-                ctype, ct_opts = werkzeug.http.parse_options_header(value)
-                boundary = ct_opts['boundary'].encode('ascii')
-                return boundary
-        return None
-
-    def __init__(self, data, boundary):
-
-        self.form = {}
-        self.files = {}
-
-        full_boundary = b'--' + boundary 
-        mid_boundary = b'\r\n' + full_boundary
-        view = memoryview(data)
-        i = data.find(full_boundary) + len(full_boundary)
-        while i >= 0:
-            next_offset = data.find(mid_boundary, i)
-            if next_offset < 0:
-                break
-            headers_offset = i + 2 # add 2 for CRLF
-            body_offset = data.find(b'\r\n\r\n', headers_offset)
-            raw_headers = view[headers_offset: body_offset]
-            body = view[body_offset + 4: next_offset]
-            i = next_offset + len(mid_boundary)
-
-            headers = {}
-            for header in raw_headers.tobytes().decode('ascii').strip().split("\r\n"):
-
-                k,v = header.split(':')
-                headers[k.strip()] = v.strip()
-
-            ctype, ct_opts = werkzeug.http.parse_options_header(headers['Content-Type'])
-            cdisp, cd_opts = werkzeug.http.parse_options_header(headers['Content-disposition'])
-            assert cdisp == 'form-data'
-
-            if 'filename' in cd_opts:
-                if cd_opts['name'] not in self.files:
-                    self.files[cd_opts['name']] = []
-
-                self.files[cd_opts['name']].append(body)
-
-            else:
-                if ctype == 'text/plain' and 'charset' in ct_opts:
-                    body = body.tobytes().decode(ct_opts['charset'])
-                if cd_opts['name'] not in self.form:
-                    self.form[cd_opts['name']] = []
-
-                self.form[cd_opts['name']].append(body)
-
-
 class DepthFormat(Enum):
     Meters = 0,
     Normalized = 1,
@@ -488,238 +364,6 @@ class Server(object):
             self.last_event.add_third_party_camera_image(img)
 
         return self.last_event
-# FifoFields
-@unique
-class FieldType(IntEnum):
-    METADATA = 1
-    ACTION = 2
-    ACTION_RESULT = 3
-    RGB_IMAGE = 4
-    DEPTH_IMAGE = 5
-    NORMALS_IMAGE = 6
-    FLOWS_IMAGE = 7
-    CLASSES_IMAGE = 8
-    IDS_IMAGE = 9
-    THIRD_PARTY_IMAGE = 10
-    END_OF_MESSAGE = 255
 
-class FifoServer(Server):
-    header_format = '!BI'
-    header_size = struct.calcsize(header_format)
-    field_types = {f.value:f for f in FieldType}
 
-    def __init__(self, width, height, depth_format=DepthFormat.Meters, add_depth_noise=False):
-
-        self.tmp_dir = tempfile.TemporaryDirectory()
-        self.server_pipe_path = os.path.join(self.tmp_dir.name, 'server.pipe')
-        self.client_pipe_path = os.path.join(self.tmp_dir.name, 'client.pipe')
-        self.server_pipe = None
-        self.client_pipe = None
-        # allows us to map the enum to form field names
-        # for backwards compatibility
-        # this can be removed when the wsgi server is removed
-        self.form_field_map = {
-            FieldType.RGB_IMAGE:'image',
-            FieldType.DEPTH_IMAGE:'image_depth',
-            FieldType.IDS_IMAGE:'image_ids',
-            FieldType.NORMALS_IMAGE:'image_normals',
-            FieldType.FLOWS_IMAGE:'image_flows',
-            FieldType.THIRD_PARTY_IMAGE:"image-thirdParty-camera"
-        }
-
-        self.image_fields = {
-            FieldType.IDS_IMAGE,
-            FieldType.CLASSES_IMAGE,
-            FieldType.FLOWS_IMAGE,
-            FieldType.NORMALS_IMAGE,
-            FieldType.DEPTH_IMAGE,
-            FieldType.RGB_IMAGE,
-            FieldType.THIRD_PARTY_IMAGE
-            }
-
-        self.eom_header = self._create_header(FieldType.END_OF_MESSAGE, b'')
-        super().__init__(width, height, depth_format, add_depth_noise)
-
-    def _create_header(self, message_type, body):
-        return struct.pack(self.header_format, message_type, len(body))
-
-    def _recv_message(self):
-        if self.server_pipe is None:
-            self.server_pipe = open(self.server_pipe_path, "rb" )
-
-        metadata = None
-        files = defaultdict(list)
-        while True:
-            header = self.server_pipe.read(self.header_size) # message type + length
-            # XXX
-            # check if len == 0 to disconnect
-            # check if pid is dead
-            
-            if header[0] == FieldType.END_OF_MESSAGE.value:
-                #print("GOT EOM")
-                break
-
-            #print("got header %s" % header)
-            field_type_int, message_length = struct.unpack(self.header_format, header)
-            field_type = self.field_types[field_type_int]
-            body = self.server_pipe.read(message_length)
-            #print("field type")
-            #print(field_type)
-            if field_type is FieldType.METADATA:
-                #print("body length %s" % len(body))
-                #print(body)
-                metadata = msgpack.loads(body, raw=False)
-                #print(metadata)
-            elif field_type in self.image_fields:
-                files[self.form_field_map[field_type]].append(body)
-            else:
-                raise ValueError("Invalid field type: %s" % field_type )
-
-        return metadata, files
-
-    def _send_message(self, message_type, body):
-        #print("trying to write to ")
-        if self.client_pipe is None:
-            self.client_pipe = open(self.client_pipe_path, "wb" )
-
-        header = self._create_header(message_type, body)
-        #print("len header %s" % len(header))
-        #print("sending body %s" % body)
-
-        self.client_pipe.write(header + body + self.eom_header)
-        self.client_pipe.flush()
-
-    def receive(self):
-        metadata, files = self._recv_message()
-        return self.create_event(metadata, files)
-
-    def send(self, action):
-        #print("got action to send")
-        if 'sequenceId' in action:
-            self.sequence_id = action['sequenceId']
-        else:
-            self.sequence_id += 1
-            action['sequenceId'] = self.sequence_id
-        #print(action)
-
-        # need to switch this to msgpack
-        self._send_message(FieldType.ACTION, json.dumps(action, cls=NumpyAwareEncoder).encode('utf8'))
-    
-
-    def start(self):
-        os.mkfifo(self.server_pipe_path)
-        os.mkfifo(self.client_pipe_path)
-        self.started = True
-
-    # params to pass up to unity
-    def unity_params(self):
-        params = dict(
-            server_type='FIFO',
-            fifo_server_pipe_path=self.server_pipe_path,
-            fifo_client_pipe_path=self.client_pipe_path
-        )
-        return params
-
-    def stop(self):
-        self.client_pipe.close()
-        self.server_pipe.close()
-
-class WsgiServer(Server):
-
-    def __init__(
-            self,
-            host,
-            port=0,
-            threaded=False,
-            depth_format=DepthFormat.Meters,
-            add_depth_noise=False,
-            width=300,
-            height=300
-    ):
-
-        app = Flask(__name__,
-                    template_folder=os.path.realpath(
-                        os.path.join(
-                            os.path.dirname(os.path.abspath(__file__)), '..', 'templates')))
-
-        self.request_queue = Queue(maxsize=1)
-        self.response_queue = Queue(maxsize=1)
-        self.app = app
-        self.app.config.update(PROPAGATE_EXCEPTIONS=False, JSONIFY_PRETTYPRINT_REGULAR=False)
-        self.port = port
-        self.last_rate_timestamp = time.time()
-        self.frame_counter = 0
-        self.debug_frames_per_interval = 50
-        self.wsgi_server = werkzeug.serving.make_server(host, self.port, self.app, threaded=threaded, request_handler=ThorRequestHandler)
-        # used to ensure that we are receiving frames for the action we sent
-        super().__init__(width, height, depth_format, add_depth_noise)
-
-        @app.route('/ping', methods=['get'])
-        def ping():
-            return 'pong'
-
-        @app.route('/train', methods=['post'])
-        def train():
-
-            if request.headers['Content-Type'].split(';')[0] == 'multipart/form-data':
-                form = MultipartFormParser(request.get_data(), MultipartFormParser.get_boundary(request.headers))
-                metadata = json.loads(form.form['metadata'][0])
-                token = form.form['token'][0]
-            else:
-                form = request
-                metadata = json.loads(form.form['metadata'])
-                token = form.form['token']
-
-            if self.client_token and token != self.client_token:
-                abort(403)
-
-            if self.frame_counter % self.debug_frames_per_interval == 0:
-                now = time.time()
-                # rate = self.debug_frames_per_interval / float(now - self.last_rate_timestamp)
-                self.last_rate_timestamp = now
-                # import datetime
-                # print("%s %s/s" % (datetime.datetime.now().isoformat(), rate))
-
-            event = self.create_event(metadata, form.files)
-
-            self.request_queue.put_nowait(event)
-
-            self.frame_counter += 1
-
-            next_action = queue_get(self.response_queue)
-            if 'sequenceId' not in next_action:
-                self.sequence_id += 1
-                next_action['sequenceId'] = self.sequence_id
-            else:
-                self.sequence_id = next_action['sequenceId']
-
-            resp = make_response(json.dumps(next_action, cls=NumpyAwareEncoder))
-
-            return resp
-
-    def _start_server_thread(self):
-        self.wsgi_server.serve_forever()
-
-    def start(self):
-        self.started = True
-        self.server_thread = threading.Thread(target=self._start_server_thread)
-        self.server_thread.daemon = True
-        self.server_thread.start()
-
-    def receive(self):
-        return queue_get(self.request_queue, self.unity_proc)
-    
-    def send(self, action):
-        assert self.request_queue.empty()
-        self.response_queue.put_nowait(action)
-
-    # params to pass up to unity
-    def unity_params(self):
-        host, port = self.wsgi_server.socket.getsockname()
-
-        params = dict(host=host, port=str(port), server_type='WSGI')
-        return params
-    
-    def stop(self):
-        self.wsgi_server.shutdown()
