+++ ai2thor/tests/test_fifo_server.py
@@ -1,8 +1,7 @@
-import ai2thor.server
+import ai2thor.fifo_server
 import pytest
 import numpy as np
 import msgpack
-from ai2thor.server import Queue
 from ai2thor.tests.test_event import metadata_simple
 from ai2thor.tests.fifo_client import FifoClient
 from io import BytesIO
@@ -14,11 +13,11 @@ import copy
 def test_multi_agent_train():
 
 
-    s = ai2thor.server.FifoServer(width=300, height=300)
+    s = ai2thor.fifo_server.FifoServer(width=300, height=300)
     s.send(dict(action='RotateRight'))
     c = FifoClient(s.server_pipe_path, s.client_pipe_path)
     msg = c.recv()
-    c.send(ai2thor.server.FieldType.METADATA, generate_multi_agent_metadata_payload(metadata_simple, s.sequence_id))
+    c.send(ai2thor.fifo_server.FieldType.METADATA, generate_multi_agent_metadata_payload(metadata_simple, s.sequence_id))
     c.send_eom()
     event = s.receive()
     assert len(event.events) == 2
@@ -28,7 +27,7 @@ def test_multi_agent_train():
 
 def test_train_numpy_action():
 
-    s = ai2thor.server.FifoServer(width=300, height=300)
+    s = ai2thor.fifo_server.FifoServer(width=300, height=300)
     s.send(dict(
         action='Teleport', 
         rotation=dict(y=np.array([24])[0]),
@@ -47,23 +46,23 @@ def generate_multi_agent_metadata_payload(metadata, sequence_id):
 
 def test_simple():
 
-    s = ai2thor.server.FifoServer(width=300, height=300)
+    s = ai2thor.fifo_server.FifoServer(width=300, height=300)
     s.send(dict(action='RotateRight'))
     c = FifoClient(s.server_pipe_path, s.client_pipe_path)
     msg = c.recv()
     assert msg == dict(action='RotateRight', sequenceId=s.sequence_id)
-    c.send(ai2thor.server.FieldType.METADATA, generate_metadata_payload(metadata_simple, s.sequence_id))
+    c.send(ai2thor.fifo_server.FieldType.METADATA, generate_metadata_payload(metadata_simple, s.sequence_id))
     c.send_eom()
     event = s.receive()
     assert event.metadata == metadata_simple
 
 def test_sequence_id_mismatch():
-    s = ai2thor.server.FifoServer(width=300, height=300)
+    s = ai2thor.fifo_server.FifoServer(width=300, height=300)
     s.send(dict(action='RotateRight'))
     c = FifoClient(s.server_pipe_path, s.client_pipe_path)
     msg = c.recv()
     assert msg == dict(action='RotateRight', sequenceId=s.sequence_id)
-    c.send(ai2thor.server.FieldType.METADATA, generate_metadata_payload(metadata_simple, s.sequence_id + 1))
+    c.send(ai2thor.fifo_server.FieldType.METADATA, generate_metadata_payload(metadata_simple, s.sequence_id + 1))
     c.send_eom()
     exception_caught = False
     try:
