+++ ai2thor/tests/test_server.py
@@ -1,8 +1,8 @@
-import ai2thor.server
+import ai2thor.wsgi_server
 import pytest
 import numpy as np
 import json
-from ai2thor.server import Queue
+from ai2thor.wsgi_server import Queue
 from ai2thor.tests.test_event import metadata_simple
 from io import BytesIO
 import copy
@@ -35,7 +35,7 @@ def generate_form(metadata, sequence_id=1):
 
 @pytest.fixture
 def server():
-    return ai2thor.server.WsgiServer(host='127.0.0.1')
+    return ai2thor.wsgi_server.WsgiServer(host='127.0.0.1')
 
 @pytest.fixture
 def client(server):
@@ -47,7 +47,7 @@ def test_ping(client):
 
 def test_multi_agent_train():
 
-    s = ai2thor.server.WsgiServer(host='127.0.0.1')
+    s = ai2thor.wsgi_server.WsgiServer(host='127.0.0.1')
     s.send(dict(action='RotateRight'))
     c = s.app.test_client()
     res = c.post(
@@ -59,7 +59,7 @@ def test_multi_agent_train():
 
 def test_train_numpy_action():
 
-    s = ai2thor.server.WsgiServer(host='127.0.0.1')
+    s = ai2thor.wsgi_server.WsgiServer(host='127.0.0.1')
     s.send(dict(
         action='Teleport', 
         rotation=dict(y=np.array([24])[0]),
@@ -77,7 +77,7 @@ def test_train_numpy_action():
 
 def test_train():
 
-    s = ai2thor.server.WsgiServer(host='127.0.0.1')
+    s = ai2thor.wsgi_server.WsgiServer(host='127.0.0.1')
     s.send(dict(action='RotateRight'))
     c = s.app.test_client()
     res = c.post(
@@ -89,7 +89,7 @@ def test_train():
 
 def test_client_token_mismatch():
 
-    s = ai2thor.server.WsgiServer(host='127.0.0.1')
+    s = ai2thor.wsgi_server.WsgiServer(host='127.0.0.1')
     s.send(dict(action='RotateRight'))
     s.client_token = '123456'
     c = s.app.test_client()
@@ -102,7 +102,7 @@ def test_client_token_mismatch():
     assert res.status_code == 403
 
 def test_non_multipart():
-    s = ai2thor.server.WsgiServer(host='127.0.0.1')
+    s = ai2thor.wsgi_server.WsgiServer(host='127.0.0.1')
     s.send(dict(action='RotateRight'))
     c = s.app.test_client()
     s.client_token = '1234567'
@@ -115,7 +115,7 @@ def test_non_multipart():
 
 def test_sequence_id_mismatch():
 
-    s = ai2thor.server.WsgiServer(host='127.0.0.1')
+    s = ai2thor.wsgi_server.WsgiServer(host='127.0.0.1')
     s.send(dict(action='RotateRight'))
     c = s.app.test_client()
